Coordinates got a class for itself.
