+++ EvoCarcassonne/Model/BoardTileModel.cs
@@ -32,22 +32,11 @@ namespace EvoCarcassonne.Model
             PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(property));
         }
 
-    }
-
-    public class Coordinates
-    {
-        public int X { get; }
-        public int Y { get; }
-
-        public Coordinates(int x, int y)
-        {
-            X = x;
-            Y = y;
-        }
-
         public override string ToString()
         {
-            return $"({ X },{ Y })";
+            return "Coordinates    " + Coordinates +
+                   "BackEndTile    " + BackendTile;
         }
     }
+
 }
