+++ EvoCarcassonne/Model/Coordinates.cs
@@ -1,0 +1,19 @@
+namespace EvoCarcassonne.Model
+{
+    public class Coordinates
+    {
+        public int X { get; }
+        public int Y { get; }
+
+        public Coordinates(int x, int y)
+        {
+            X = x;
+            Y = y;
+        }
+
+        public override string ToString()
+        {
+            return $"({ X },{ Y })";
+        }
+    }
+}
